+++ project/src/components/review-item/review-item.tsx
@@ -1,4 +1,4 @@
-import { ReviewType } from '../../types/offer';
+import { ReviewType } from '../../types/review';
 
 type ReviewItemProps = {
   review: ReviewType;
