+++ project/src/const.ts
@@ -15,3 +15,12 @@ export enum AutorizationStatus {
 
 export const URL_MARKER_DEFAULT = '../public/img/pin.svg';
 export const URL_MARKER_CURRENT = '../public/img/pin-active.svg';
+
+export const CitiesList = [
+  'Paris',
+  'Cologne',
+  'Brussels',
+  'Amsterdam',
+  'Hamburg',
+  'Dusseldorf',
+];
