+++ project/src/pages/room-screen.tsx
@@ -1,0 +1,257 @@
+import { Link } from 'react-router-dom';
+import Header from '../components/header/header';
+import { OfferType } from '../types/offer';
+import { ReviewType } from '../types/review';
+import getPercRating from '../utils';
+import ReviewForm from '../components/review-form/review-form';
+import ReviewList from '../components/review-list/review-list';
+
+type RoomProps={
+  offers: OfferType;
+  reviews: ReviewType[];
+}
+
+function Room({offers, reviews}: RoomProps): JSX.Element {
+  const {rating, type, bedrooms, maxAdults, description, price, isPremium} = offers;
+  return (
+    <>
+      <Header/>
+      <main className="page__main page__main--property">
+        <section className="property">
+          <div className="property__gallery-container container">
+            <div className="property__gallery">
+              <div className="property__image-wrapper">
+                <img className="property__image" src="img/room.jpg" alt="studio" />
+              </div>
+              <div className="property__image-wrapper">
+                <img className="property__image" src="img/apartment-01.jpg" alt="studio" />
+              </div>
+              <div className="property__image-wrapper">
+                <img className="property__image" src="img/apartment-02.jpg" alt="studio" />
+              </div>
+              <div className="property__image-wrapper">
+                <img className="property__image" src="img/apartment-03.jpg" alt="studio" />
+              </div>
+              <div className="property__image-wrapper">
+                <img className="property__image" src="img/studio-01.jpg" alt="studio" />
+              </div>
+              <div className="property__image-wrapper">
+                <img className="property__image" src="img/apartment-01.jpg" alt="studio" />
+              </div>
+            </div>
+          </div>
+          <div className="property__container container">
+            <div className="property__wrapper">
+              {
+                (isPremium) &&
+                <div className="property__mark">
+                  <span>Premium</span>
+                </div>
+              }
+              <div className="property__name-wrapper">
+                <h1 className="property__name">
+                    Beautiful &amp; luxurious studio at great location
+                </h1>
+                <button className="property__bookmark-button button" type="button">
+                  <svg className="property__bookmark-icon" width="31" height="33">
+                    <use xlinkHref="#icon-bookmark"></use>
+                  </svg>
+                  <span className="visually-hidden">To bookmarks</span>
+                </button>
+              </div>
+              <div className="property__rating rating">
+                <div className="property__stars rating__stars">
+                  <span style={{width: `${getPercRating(rating)}%`}}></span>
+                  <span className="visually-hidden">Rating</span>
+                </div>
+                <span className="property__rating-value rating__value">{rating}</span>
+              </div>
+              <ul className="property__features">
+                <li className="property__feature property__feature--entire">
+                  {type}
+                </li>
+                <li className="property__feature property__feature--bedrooms">
+                  {bedrooms}
+                </li>
+                <li className="property__feature property__feature--adults">
+                  {maxAdults}
+                </li>
+              </ul>
+              <div className="property__price">
+                <b className="property__price-value">&euro;{price}</b>
+                <span className="property__price-text">&nbsp;night</span>
+              </div>
+              <div className="property__inside">
+                <h2 className="property__inside-title">What&apos;s inside</h2>
+                <ul className="property__inside-list">
+                  <li className="property__inside-item">
+                      Wi-Fi
+                  </li>
+                  <li className="property__inside-item">
+                      Washing machine
+                  </li>
+                  <li className="property__inside-item">
+                      Towels
+                  </li>
+                  <li className="property__inside-item">
+                      Heating
+                  </li>
+                  <li className="property__inside-item">
+                      Coffee machine
+                  </li>
+                  <li className="property__inside-item">
+                      Baby seat
+                  </li>
+                  <li className="property__inside-item">
+                      Kitchen
+                  </li>
+                  <li className="property__inside-item">
+                      Dishwasher
+                  </li>
+                  <li className="property__inside-item">
+                      Cabel TV
+                  </li>
+                  <li className="property__inside-item">
+                      Fridge
+                  </li>
+                </ul>
+              </div>
+              <div className="property__host">
+                <h2 className="property__host-title">Meet the host</h2>
+                <div className="property__host-user user">
+                  <div className="property__avatar-wrapper property__avatar-wrapper--pro user__avatar-wrapper">
+                    <img className="property__avatar user__avatar" src="img/avatar-angelina.jpg" width="74" height="74" alt="Host avatar" />
+                  </div>
+                  <span className="property__user-name">
+                      Angelina
+                  </span>
+                  <span className="property__user-status">
+                      Pro
+                  </span>
+                </div>
+                <div className="property__description">
+                  <p className="property__text">
+                    {description}
+                  </p>
+                </div>
+              </div>
+              <section className="property__reviews reviews">
+                <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">1</span></h2>
+                <ReviewList reviews={reviews}/>
+                <ReviewForm/>
+              </section>
+            </div>
+          </div>
+          <section className="property__map map"></section>
+        </section>
+        <div className="container">
+          <section className="near-places places">
+            <h2 className="near-places__title">Other places in the neighbourhood</h2>
+            <div className="near-places__list places__list">
+              <article className="near-places__card place-card">
+                <div className="near-places__image-wrapper place-card__image-wrapper">
+                  <Link to=" /">
+                    <img className="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place" />
+                  </Link>
+                </div>
+                <div className="place-card__info">
+                  <div className="place-card__price-wrapper">
+                    <div className="place-card__price">
+                      <b className="place-card__price-value">&euro;80</b>
+                      <span className="place-card__price-text">&#47;&nbsp;night</span>
+                    </div>
+                    <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
+                      <svg className="place-card__bookmark-icon" width="18" height="19">
+                        <use xlinkHref="#icon-bookmark"></use>
+                      </svg>
+                      <span className="visually-hidden">In bookmarks</span>
+                    </button>
+                  </div>
+                  <div className="place-card__rating rating">
+                    <div className="place-card__stars rating__stars">
+                      <span style={{width: 80}}></span>
+                      <span className="visually-hidden">Rating</span>
+                    </div>
+                  </div>
+                  <h2 className="place-card__name">
+                    <Link to=" /">Wood and stone place</Link>
+                  </h2>
+                  <p className="place-card__type">Private room</p>
+                </div>
+              </article>
+
+              <article className="near-places__card place-card">
+                <div className="near-places__image-wrapper place-card__image-wrapper">
+                  <a href=" ">
+                    <img className="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place" />
+                  </a>
+                </div>
+                <div className="place-card__info">
+                  <div className="place-card__price-wrapper">
+                    <div className="place-card__price">
+                      <b className="place-card__price-value">&euro;132</b>
+                      <span className="place-card__price-text">&#47;&nbsp;night</span>
+                    </div>
+                    <button className="place-card__bookmark-button button" type="button">
+                      <svg className="place-card__bookmark-icon" width="18" height="19">
+                        <use xlinkHref="#icon-bookmark"></use>
+                      </svg>
+                      <span className="visually-hidden">To bookmarks</span>
+                    </button>
+                  </div>
+                  <div className="place-card__rating rating">
+                    <div className="place-card__stars rating__stars">
+                      <span style={{width: 80}}></span>
+                      <span className="visually-hidden">Rating</span>
+                    </div>
+                  </div>
+                  <h2 className="place-card__name">
+                    <a href=" ">Canal View Prinsengracht</a>
+                  </h2>
+                  <p className="place-card__type">Apartment</p>
+                </div>
+              </article>
+
+              <article className="near-places__card place-card">
+                <div className="place-card__mark">
+                  <span>Premium</span>
+                </div>
+                <div className="near-places__image-wrapper place-card__image-wrapper">
+                  <a href=" ">
+                    <img className="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place" />
+                  </a>
+                </div>
+                <div className="place-card__info">
+                  <div className="place-card__price-wrapper">
+                    <div className="place-card__price">
+                      <b className="place-card__price-value">&euro;180</b>
+                      <span className="place-card__price-text">&#47;&nbsp;night</span>
+                    </div>
+                    <button className="place-card__bookmark-button button" type="button">
+                      <svg className="place-card__bookmark-icon" width="18" height="19">
+                        <use xlinkHref="#icon-bookmark"></use>
+                      </svg>
+                      <span className="visually-hidden">To bookmarks</span>
+                    </button>
+                  </div>
+                  <div className="place-card__rating rating">
+                    <div className="place-card__stars rating__stars">
+                      <span style={{width: 100}}></span>
+                      <span className="visually-hidden">Rating</span>
+                    </div>
+                  </div>
+                  <h2 className="place-card__name">
+                    <a href=" ">Nice, cozy, warm big bed apartment</a>
+                  </h2>
+                  <p className="place-card__type">Apartment</p>
+                </div>
+              </article>
+            </div>
+          </section>
+        </div>
+      </main>
+    </>
+  );
+}
+
+export default Room;
