+++ project/src/pages/sign-in.tsx
@@ -1,5 +1,5 @@
 import {Link} from 'react-router-dom';
-import Header from '../header/header';
+import Header from '../components/header/header';
 
 function SignIn(): JSX.Element {
   return (
