+++ project/src/types/city.ts
@@ -1,0 +1,6 @@
+export type City = {
+  title: string;
+  lat: number;
+  lng: number;
+  zoom: number;
+};
