+++ project/src/types/favorite-offer.ts
@@ -1,0 +1,26 @@
+export type FavoriteOfferType= {
+  bedrooms: number;
+  city: {
+    location: Location,
+    name: string;
+  }
+  description: string;
+  goods: string[];
+  host: {
+    avatarUrl: string;
+    id: number;
+    isPro: boolean;
+    name: string;
+  }
+  id: number;
+  images: string[];
+  isFavorite: boolean;
+  isPremium: boolean;
+  location: Location;
+  maxAdults: number;
+  previewImage: string;
+  price: number;
+  rating: number;
+  title: string;
+  type: string;
+};
