+++ project/src/types/offer.ts
@@ -1,11 +1,13 @@
+export type Location = {
+  latitude: number
+  longitude: number
+  zoom: number
+}
+
 export type OfferType = {
   bedrooms: number,
   city: {
-    location: {
-      latitude: number
-      longitude: number
-      zoom: number
-    },
+    location: Location,
     name: string
   },
   description: string,
@@ -20,70 +22,11 @@ export type OfferType = {
   images: string[],
   isFavorite: boolean,
   isPremium: boolean,
-  location: {
-    latitude: number
-    longitude: number
-    zoom: number
-  },
+  location: Location,
   maxAdults: number,
   previewImage: string,
   price: number,
   rating: number,
   title: string,
   type: string
-};
-
-export type ReviewType = {
-  comment: string
-  date: string
-  id: number
-  rating: number
-  user: {
-    avatarUrl: string
-    id: number
-    isPro: boolean
-    name: string
-  }
-};
-
-export type FavoriteOfferType= {
-  bedrooms: number;
-  city: {
-    location: {
-      latitude: number;
-      longitude: number;
-      zoom: number;
-    }
-    name: string;
-  }
-  description: string;
-  goods: string[];
-  host: {
-    avatarUrl: string;
-    id: number;
-    isPro: boolean;
-    name: string;
-  }
-  id: number;
-  images: string[];
-  isFavorite: boolean;
-  isPremium: boolean;
-  location: {
-    latitude: number;
-    longitude: number;
-    zoom: number;
-  }
-  maxAdults: number;
-  previewImage: string;
-  price: number;
-  rating: number;
-  title: string;
-  type: string;
-};
-
-export type City = {
-  title: string;
-  lat: number;
-  lng: number;
-  zoom: number;
 };
