+++ project/src/types/review.ts
@@ -1,0 +1,12 @@
+export type ReviewType = {
+  comment: string
+  date: string
+  id: number
+  rating: number
+  user: {
+    avatarUrl: string
+    id: number
+    isPro: boolean
+    name: string
+  }
+};
